+++ wcfsetup/install/files/lib/data/TMultiCategoryObject.class.php
@@ -53,7 +53,7 @@ trait TMultiCategoryObject {
 	/**
 	 * Returns the categories of the object.
 	 *
-	 * @return	CalendarCategory[]
+	 * @return	AbstractDecoratedCategory[]
 	 */
 	public function getCategories() {
 		if ($this->categories === null) {